requirements.txt
streamlit
langchain-community
langchain-text-splitters
langchain-google-genai
google-generativeai
chromadb
pypdf
python-dotenv
